Fixed a bunch of Xcode warnings.

+++ src/realm/impl/transact_log.cpp
@@ -66,7 +66,7 @@ void TransactLogConvenientEncoder::do_select_table(const Table* table)
     }
     std::reverse(begin, end);
 
-    int levels = (end - begin) / 2;
+    size_t levels = (end - begin) / 2;
     m_encoder.select_table(*begin, levels, begin + 1); // Throws
     m_selected_spec = nullptr;
     m_selected_link_list = nullptr;
